[Gather] update kernel interface
