+++ src/ops/kernels/gather_kernels.cu
@@ -14,6 +14,7 @@
  */
 
 #include "flexflow/ops/gather.h"
+#include "flexflow/ops/kernels/gather_kernels.h"
 #include "flexflow/utils/cuda_helper.h"
 
 namespace FlexFlow {
@@ -26,10 +27,13 @@ GatherMeta::GatherMeta(FFHandler handler, Gather const *gather)
   legion_dim = gather->legion_dim;
 }
 
-void Gather::forward_kernel_wrapper(GatherMeta const *m,
-                                    GenericTensorAccessorR const &input,
-                                    GenericTensorAccessorR const &index,
-                                    GenericTensorAccessorW const &output) {
+namespace Kernels {
+namespace Gather {
+
+void forward_kernel_wrapper(GatherMeta const *m,
+                            GenericTensorAccessorR const &input,
+                            GenericTensorAccessorR const &index,
+                            GenericTensorAccessorW const &output) {
   cudaStream_t stream;
   checkCUDA(get_legion_stream(&stream));
   coord_t stride = 1;
@@ -39,24 +43,58 @@ void Gather::forward_kernel_wrapper(GatherMeta const *m,
   coord_t dim_size =
       output.domain.hi()[m->legion_dim] - output.domain.lo()[m->legion_dim] + 1;
   if (index.data_type == DT_INT32) {
-    Gather::forward_kernel(input.get_float_ptr(),
-                           index.get_int32_ptr(),
-                           output.get_float_ptr(),
-                           output.domain.get_volume(),
-                           stride,
-                           dim_size,
-                           stream);
+    Internal::forward_kernel(input.get_float_ptr(),
+                             index.get_int32_ptr(),
+                             output.get_float_ptr(),
+                             output.domain.get_volume(),
+                             stride,
+                             dim_size,
+                             stream);
   } else {
     assert(index.data_type == DT_INT64);
-    Gather::forward_kernel(input.get_float_ptr(),
-                           index.get_int64_ptr(),
-                           output.get_float_ptr(),
-                           output.domain.get_volume(),
-                           stride,
-                           dim_size,
-                           stream);
+    Internal::forward_kernel(input.get_float_ptr(),
+                             index.get_int64_ptr(),
+                             output.get_float_ptr(),
+                             output.domain.get_volume(),
+                             stride,
+                             dim_size,
+                             stream);
   }
 }
+
+void backward_kernel_wrapper(GatherMeta const *m,
+                             GenericTensorAccessorR const &output_grad,
+                             GenericTensorAccessorR const &index,
+                             GenericTensorAccessorW const &input_grad) {
+  cudaStream_t stream;
+  checkCUDA(get_legion_stream(&stream));
+  coord_t stride = 1;
+  for (int i = 0; i < m->legion_dim; i++) {
+    stride *= (output_grad.domain.hi()[i] - output_grad.domain.lo()[i] + 1);
+  }
+  coord_t dim_size = output_grad.domain.hi()[m->legion_dim] -
+                     output_grad.domain.lo()[m->legion_dim] + 1;
+  if (index.data_type == DT_INT32) {
+    Internal::backward_kernel(output_grad.get_float_ptr(),
+                              index.get_int32_ptr(),
+                              input_grad.get_float_ptr(),
+                              output_grad.domain.get_volume(),
+                              stride,
+                              dim_size,
+                              stream);
+  } else {
+    assert(index.data_type == DT_INT64);
+    Internal::backward_kernel(output_grad.get_float_ptr(),
+                              index.get_int64_ptr(),
+                              input_grad.get_float_ptr(),
+                              output_grad.domain.get_volume(),
+                              stride,
+                              dim_size,
+                              stream);
+  }
+}
+
+namespace Internal {
 
 template <typename IndexType>
 __global__ void gather_forward(float const *input,
@@ -79,51 +117,19 @@ __global__ void gather_forward(float const *input,
 }
 
 template <typename IndexType>
-void Gather::forward_kernel(float const *input_ptr,
-                            IndexType const *index_ptr,
-                            float *output_ptr,
-                            coord_t output_size,
-                            coord_t stride,
-                            coord_t dim_size,
-                            cudaStream_t stream) {
+void forward_kernel(float const *input_ptr,
+                    IndexType const *index_ptr,
+                    float *output_ptr,
+                    coord_t output_size,
+                    coord_t stride,
+                    coord_t dim_size,
+                    cudaStream_t stream) {
   assert(input_ptr != nullptr);
   assert(index_ptr != nullptr);
   assert(output_ptr != nullptr);
   gather_forward<IndexType>
       <<<GET_BLOCKS(output_size), CUDA_NUM_THREADS, 0, stream>>>(
           input_ptr, index_ptr, output_ptr, output_size, stride, dim_size);
-}
-
-void Gather::backward_kernel_wrapper(GatherMeta const *m,
-                                     GenericTensorAccessorR const &output_grad,
-                                     GenericTensorAccessorR const &index,
-                                     GenericTensorAccessorW const &input_grad) {
-  cudaStream_t stream;
-  checkCUDA(get_legion_stream(&stream));
-  coord_t stride = 1;
-  for (int i = 0; i < m->legion_dim; i++) {
-    stride *= (output_grad.domain.hi()[i] - output_grad.domain.lo()[i] + 1);
-  }
-  coord_t dim_size = output_grad.domain.hi()[m->legion_dim] -
-                     output_grad.domain.lo()[m->legion_dim] + 1;
-  if (index.data_type == DT_INT32) {
-    Gather::backward_kernel(output_grad.get_float_ptr(),
-                            index.get_int32_ptr(),
-                            input_grad.get_float_ptr(),
-                            output_grad.domain.get_volume(),
-                            stride,
-                            dim_size,
-                            stream);
-  } else {
-    assert(index.data_type == DT_INT64);
-    Gather::backward_kernel(output_grad.get_float_ptr(),
-                            index.get_int64_ptr(),
-                            input_grad.get_float_ptr(),
-                            output_grad.domain.get_volume(),
-                            stride,
-                            dim_size,
-                            stream);
-  }
 }
 
 template <typename IndexType>
@@ -147,13 +153,13 @@ __global__ void gather_backward(float const *output_grad,
 }
 
 template <typename IndexType>
-void Gather::backward_kernel(float const *output_grad_ptr,
-                             IndexType const *index_ptr,
-                             float *input_grad_ptr,
-                             coord_t output_size,
-                             coord_t stride,
-                             coord_t dim_size,
-                             cudaStream_t stream) {
+void backward_kernel(float const *output_grad_ptr,
+                     IndexType const *index_ptr,
+                     float *input_grad_ptr,
+                     coord_t output_size,
+                     coord_t stride,
+                     coord_t dim_size,
+                     cudaStream_t stream) {
   assert(output_grad_ptr != nullptr);
   assert(input_grad_ptr != nullptr);
   assert(index_ptr != nullptr);
@@ -166,5 +172,9 @@ void Gather::backward_kernel(float const *output_grad_ptr,
           stride,
           dim_size);
 }
+
+} // namespace Internal
+} // namespace Gather
+} // namespace Kernels
 
 }; // namespace FlexFlow
